requirements.txt
djongo==1.3.6
dnspython 
pymongo==3.12.3
Pillow 
pytz
django-shopping-cart
gunicorn
razorpay 
sqlparse
psycopg2
psycopg2-binary